+++ crates/igp-oracle/src/cosmosnative/mod.rs
@@ -1,1 +1,2 @@
+pub mod adapter;
 pub mod query;
